feat: multi resolution adaptation (#171)
## Description

The adaptation for the multi-resolution used a copy of the previous
field. This field is used when, during the adaptation, we have the
following phenomena: at iteration `i` of the adaptation algorithm, a
cell is made coarser; the at iteration `i+1` the coarser cell is refined
going back to the previous state. We save the field to not make a
prediction of these refine cells, but just put their previous values.

The algorithm defined by Harten is recursive and the behavior described
above should not occur. To remove the copy of the previous field, we
have to compute two level of details. In this way, the first detail will
tell us to make the cell coarserl and the next one to refine it, leaving
the cell unchanged. This is what is done in this PR.


## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct