fix: periodic (#188)
## Description
The update ghost for the periodic case is broken. We update the ghosts
by looking for the cell strip on the other side. Unfortunately, there
may be cases where there are no corresponding ghosts. So it's best to
take the ghost strip you're interested in and search for it inside the
domain on the other side. This PR fixes this issue by changing the
search.

## Related issue
The case `linear_convection.cpp` in `demos/FiniteVolume`was broken.

## How has this been tested?
`linear_convection.cpp` tests the validity of this PR.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct