build: proper link to PETSc library with absolute path
When using `pkg_search_module` to find PETSc, it populates (at least) two
variables: `PETSC_LIBRARIES` that contains only the name of the library
(ie `petsc`) and `PETSC_LINK_LIBRARIES` that contains the absolute path
to the library.

Until this fix, only `PETSC_LIBRARIES` was used that leads to linkage
error for some configurations, eg when PETSc is not installed system-wide
but only through a package manager like in a conda environment.

Using `PETSC_LINK_LIBRARIES` instead fix this issue.