feat: Scaling (#229)
## Description
A scaling factor is taken into account so that the box [0,1]^d and the
box [a,b]^d, for any a,b, offer the same levels of refinement. In the
case of a square box, level 0 now corresponds to one single cell
covering the whole domain. In the case of a rectangular box, the cell
length at level 0 is dynamically computed to ensure that the box's shape
is correctly represented up to a given tolerance. This means that level
0 can already be pretty refined.

#### BREAKING CHANGES

- To recover the same refinement levels, `min_level` and `max_level`
must be updated. For instance, for the box [-1, 1]^d, the former level L
now corresponds to L+1. min and max-level must therefore be incremented
by 1 to recover the same simulation. On the other hand, if the box is,
say, [-3, 3], there is no way to recover exactly the same levels as
before, since the length of the domain (6) is not a power of 2. Indeed,
before, the cells were of length 1 at level 0, 2 at level 1, etc. Now,
of length 6 at level 0, 3 at level 1, 1.5 at level 2, etc.
- The function calls to the function `samurai::cell_length(level)` will
no longer compile. All occurrences must me replaced with
`mesh.cell_length(level)`.

## Related issue

1. All levels were according to the [0,1] interval, such that the
interval domain [0, 1000] yielded 1000 cells at level 0 without any
possibility of generating coarser cells.
2. Rectangular domains provoked bugs when they couldn't be represented
in the discrete setting. For instance, the domain changed shapes
according to the level of refinement and therefore could take weird
non-squared shapes on adapted meshes.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct