perf: remove apply interval index (#242)
## Description
This PR removes the computation of the index of each interval found in
the algebra of set. This index is only used during the renumbering phase
of the sub meshes. We improve the performance of the renumbering by
removing the old implementation.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct