feat: new container types (#222)
## Description
This PR adds the possibility to change the container used to store the
values of a field. Two types of container are available for now:

- xtensor
- eigen

It's also possible to change the memory layout (row major or column
major). There are still various bugs in Eigen due to some
incompatibilities between row major and column major when we pass from a
vector to a scalar. This is often the case when using the views and the
lazy evaluation. This issue should be investigated in a future work.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct

---------

Co-authored-by: Pierre Matalon <pierre.matalon@gmail.com>