feat: new set algebra (#275)
## Description

Currently, the algebra of sets written in Samurai doesn’t provide all
the features needed to write concise code. One of the main problems is
the impossibility to translate a set or to define a different projection
level for each set in an expression. For example, all these expressions
are not possible with the current implementation:

```cpp
auto set1 = translate(intersection(A, B), {1, 1});
auto set2 = intersection(A.on(3), B);
auto set3 = intersection(A, union_(B.on(2), C).on(5)).on(7);
```

This PR implements a complete algebra of set, which will reduce some
part of the current code.

## How has this been tested?

check the file `tests/test_subset.cpp`

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct