chore: update pre-commit (#276)
## Description
This PR update the version of the tools used in pre-commit. The version
of clang-format is updated to v1.18.2 in order to have the same behavior
with vscode when we format "on save" using the clang-format on our
system.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct