feat: min and max level can be changed (#277)
## Description
When a mesh is created, the minimum and maximum levels can only be read.
It can be interesting to be able to change these values during the
computation. This PR adds a set method for both.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct