feat: restart (#278)
This PR implements the restart feature.

- see `./tests/test_restart.cpp`
- Most of the finite volume demos also have an example of how to use the
restart. Three of them are tested during the CI: we restart from the
init file and we check if the result is the same as the reference at the
end of the computation.

By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct