perf: Naive neighbourhood (#318)
## Description
Here is a naive solution to construct in a naive way the neighbourhood. 
It is optimal as long as you are using a naive partition as for now. 
Each rank have one or two neighbours. 
This do not works properly with periodic.
We need to improve this implementation by adding new
`find_neighbourhood_<strategy>` in the future.

## Related issue
Perforance issue

## How has this been tested?
on advection-2d

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct