feat: parallel restart (#331)
## Description
This PR allows a parallel simulation to be restarted. Since load
balancing has not yet been completed, the restart process checks that
the number of processes is the same in the simulation that is being
restarted and in the restart file. If not, an error message is
displayed.

This constraint will be removed once load balancing is working as
expected.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct