feat: check if a set is empty (#332)
## Description
For now, there is a method in set algebra called 'exist' that checks
whether a subset can exist. For example, the intersection of a set and
an empty set is also empty, so there is no need to go further. The same
applies to the difference between an empty set and a set.

However, it may also be interesting to know whether the set can exist
and whether it is empty. This PR implements this method. If we find an
interval, we return true immediately. Otherwise, we search through the
set until the end.

## How has this been tested?
A test has been added in `test_subset.cpp` file.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct