feat: add finer-level-flux option (#359)
## Description
This PR removes the option `--enable--max-level-flux` and replaces it
with the more flexible option `--finer-level-flux`. Setting this option
to `-1` results in the same behaviour as before with the option
`--enable--max-level-flux`. If a positive integer `i` is provided, the
flux is computed at `level + i`.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct