fix: h_face (#371)
## Description
The computation of the face length when the portion was used was
incorrect. It takes the `max_level` of the cells, rather than the global
`max_level` of the simulation. Therefore, the scaling factor was
incorrect, producing a solution that moved too fast.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct